move-components-folder-out-pages
