+++ client/src/components/ContentBox.tsx
