+++ client/src/components/HomeProjects.tsx
