+++ client/src/components/SocialLinks.tsx
@@ -1,0 +1,10 @@
+ const socialLinks = {
+    instagram: "#", 
+    facebook: "https://www.facebook.com/dINChiII",
+    twitter: "https://twitter.com/DincherSel32807",
+    github: "https://github.com/dincherselimov",
+    linkedin: "https://www.linkedin.com/in/dincher-selimov-6b069b209/",
+  };
+  
+
+  export default socialLinks;
