+++ client/src/pages/posts/Layout.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import Header from "../components/Header";
+import Header from "../../components/Header";
 // import Footer from "../components/Footer";
 
 export default function PageLayout({ children }: { children: ReactNode }) {
@@ -7,7 +7,6 @@ export default function PageLayout({ children }: { children: ReactNode }) {
     <div>
       <Header/>
        <main>{children}</main>
-      {/* <Footer /> */}
     </div>
   );
 }
