+++ client/src/styles/home/form.css
@@ -110,3 +110,18 @@
         font-size:18px;
     }
   }
+
+  @media only screen and (min-width: 768px) and (max-width: 1400px) {
+    .form-container {
+        padding: 40px;
+        border-radius: 10px;
+        box-shadow: 0px 0px 10px 10px rgba(219, 171, 6, 0.62);
+        max-width: 600px;
+        width: 50%;
+        text-align: center;
+        margin-left:30px;
+    }
+    .reach-out{
+        font-size:22px;
+    }
+  }
