+++ client/src/styles/home/motivation.css
@@ -1,7 +1,8 @@
 /* Motivation.css */
 
 .motivation-section {
-    background-color: #fcc200d9; /* Set your desired background color */
+    background-color: #fbc100;
+    background-image: linear-gradient(45deg, #fbc100 51%, #c46443 73%);
     padding: 20px;
     text-align: center;
 }
