+++ client/src/styles/home/welcome.css
@@ -6,6 +6,14 @@
   color: #1e205a;
 }
 
+.welcomeSection h1 {
+  background-size: 200% auto;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-color: #fbc100;
+  background-image: linear-gradient(270deg, #fbc100 51%, #c46443 76%);
+}
+
 .social-links {
   display: flex;
 }
@@ -64,6 +72,76 @@
 }
 
 
+@media only screen and (min-width:768px){
+  .welcomeSection h1 {
+    color: white;
+    font-size: 32px;
+    padding-bottom: 20px;
+    font-weight: 600;
+    width:70%;
+  }
+  .social-links {
+    display: flex;
+    margin-left: -30px;
+  }
+  
+  .social-links a {
+    width: 80px;
+    height: 80px;
+    text-align: center;
+    text-decoration: none;
+    color: #000;
+    box-shadow: 0 0 50px 10px rgba(0, 0, 0, 0.1);
+    margin: 0 0 0 30px;
+    border-radius: 50%;
+    position: relative;
+    overflow: hidden;
+    transition: transform 0.5s;
+    background: #fcc200;
+  }
+  
+  .social-links a .fab {
+    font-size: 20px;
+    line-height: 80px;
+    position: relative;
+    z-index: 10;
+    transition: color 0.5s;
+  }
+  
+  .social-links a::after {
+    content: '';
+    width: 100%;
+    height: 100%;
+    top: -90px;
+    left: 0;
+    background: #000;
+    background: linear-gradient(to top, #000046, #1cb5e0); 
+    position: absolute;
+    transition: 0.5s;
+  }
+  
+  .social-links a:hover::after {
+    top: 0;
+  }
+  
+  .social-links a:hover .fab {
+    color: #fff;
+  }
+  
+  .social-links a:hover {
+    transform: translateY(-10px);
+  }
+  .me {
+    border-radius: 10%;
+    width: 300px;
+    height: 100%;
+    background: #fbfbfbab;
+  }
+  .dinko{
+    padding:30px;
+  }
+}
+
 
 @media screen and (min-width: 992px) and (max-width: 1400px) {
   .welcomeSection {
@@ -71,11 +149,13 @@
     height: 25rem;
     display: flex;
     align-items: center;
+    background: #48494433;
+    border-radius: 40px;
    
   }
 
   .welcomeSection h1 {
-    color: #0e2355;
+    color: white;
     font-size: 32px;
     padding-bottom: 20px;
     font-weight: 600;
@@ -163,6 +243,8 @@
 
 
 
+
+
 @media only screen and (min-width:1400px){
   .welcomeSection {
     width: 100%;
@@ -171,6 +253,8 @@
     align-items: center;
     background: #48494433;
     border-radius: 40px;
+  
+
   }
  
   .welcomeSection h1 {
@@ -178,6 +262,7 @@
     font-size: 32px;
     padding-bottom:20px;
     font-weight:600;
+    width:100%;
   }
   .container {
     max-width: 1300px;
